ai-data-cleaner/requirements.txt
# Core FastAPI and web server dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Data processing
pandas>=2.2.0
numpy>=1.26.0
pyspark>=3.5.0

# Google Generative AI (Gemini)
google-generativeai>=0.8.0

# Visualization
matplotlib>=3.9.0
seaborn>=0.13.2

# File format support
pyarrow>=17.0.0

# Logging and utilities
python-dotenv>=1.0.1

# Additional dependencies for Python 3.13 compatibility
typing-extensions>=4.12.0

# Development dependencies (optional)
pytest>=8.3.0
pytest-asyncio>=0.24.0
black>=24.10.0
flake8>=7.1.0
